requirements.txt
# Core Libraries
pandas
numpy
scikit-learn
imbalanced-learn
joblib

# GPU-Accelerated Libraries
xgboost
cuml
cupy

# Colab & Cloud Storage
google-colab

# Data Visualization (if used in exploration)
matplotlib
seaborn




